Messed with contact page
I accidentally put all the content for the contact page on the services page, I pasted my work to the correct page but its still on the services page. I am trying to get the photos to site next to the lines of text but its not going great so far, I'm gonna look into it!