Played with SEO
Added favicon to template and added that to the headers (not sure if it did much, I'm trying to see if it improves SEO score). I also added a meta description to each of the pages. I cleaned up the alt text in the footer and on the services and contact page